Implement filter support for the List Objects command
Support for filtering returned objects by their type
